tools: Update magiskboot to fix generation of garbage boot images
This is magiskboot with [1] applied. The reason the binary is larger than
before is because I couldn't get magiskboot to compile with LTO, because
magiskboot utilizes several Rust feature flags that are only permitted in a
nightly Rust toolchain release, and I don't know which nightly Rust version
is intended to be used to compile magiskboot; when I compile magiskboot
with the latest Rust nightly, LLD emits an error at link-time with LTO
enabled, hence I disabled LTO.

With this update, magiskboot will no longer generate garbage boot images
when this AK3 zip is sideloaded in a recovery.

[1] https://github.com/topjohnwu/Magisk/pull/8368

Signed-off-by: Cyber Knight <cyberknight755@gmail.com>